+++ Release-Notes.md
@@ -1,5 +1,14 @@
 # FarNet.SQLite Release Notes
 
+## v0.1.0
+
+Command parameters: `IDictionary` or `SQLiteParameter` for named parameters or objects for positional parameters.
+Thus:
+
+- `IDictionary` works as before.
+- Positional parameters are supported.
+- Tuples as parameters are not supported.
+
 ## v0.0.2
 
 System.Data.SQLite 1.0.117, SQLite 3.40.0